Refactoring some tests.
