+++ testing/test_awswrangler/test_db.py
@@ -9,7 +9,13 @@ import sqlalchemy
 
 import awswrangler as wr
 
-from ._utils import CFN_VALID_STATUS, ensure_data_types, ensure_data_types_category, get_df, get_df_category
+from ._utils import (
+    ensure_data_types,
+    ensure_data_types_category,
+    extract_cloudformation_outputs,
+    get_df,
+    get_df_category,
+)
 
 logging.basicConfig(level=logging.INFO, format="[%(asctime)s][%(levelname)s][%(name)s][%(funcName)s] %(message)s")
 logging.getLogger("awswrangler").setLevel(logging.DEBUG)
@@ -18,12 +24,7 @@ logging.getLogger("botocore.credentials").setLevel(logging.CRITICAL)
 
 @pytest.fixture(scope="module")
 def cloudformation_outputs():
-    response = boto3.client("cloudformation").describe_stacks(StackName="aws-data-wrangler")
-    stack = [x for x in response.get("Stacks") if x["StackStatus"] in CFN_VALID_STATUS][0]
-    outputs = {}
-    for output in stack.get("Outputs"):
-        outputs[output.get("OutputKey")] = output.get("OutputValue")
-    yield outputs
+    yield extract_cloudformation_outputs()
 
 
 @pytest.fixture(scope="module")
